improved precision of time parsing git-svn-id: https://svn-agbkb.informatik.uni-bremen.de/Hets/trunk@7994 cec4b9c1-7d33-0410-9eda-942365e851bb

+++ Propositional/Prove.hs
@@ -419,7 +419,7 @@ analyzeZchaff str pState =
 calculateTime :: String -> TimeOfDay
 calculateTime timeLine = 
     timeToTimeOfDay $ realToFrac $ ((read $ subRegex re_SUBPOINT 
-               (subRegex re_SUBTIME timeLine "") "")::Float)
+               (subRegex re_SUBTIME timeLine "") "")::Double)
       
 re_UNSAT :: Regex  
 re_UNSAT = mkRegex "(.*)RESULT:UNSAT(.*)"
